Switch to usual setup
